bevy_core_pipeline::auto_exposure
Type Alias AutoExposureSettings

pub type AutoExposureSettings = AutoExposure;

ğŸ‘Deprecated since 0.15.0: Renamed to AutoExposure

Aliased Type

struct AutoExposureSettings {
    pub range: RangeInclusive<f32>,
    pub filter: RangeInclusive<f32>,
    pub speed_brighten: f32,
    pub speed_darken: f32,
    pub exponential_transition_distance: f32,
    pub metering_mask: Handle<Image>,
    pub compensation_curve: Handle<AutoExposureCompensationCurve>,
}

Fields
range: RangeInclusive<f32>

The range of exposure values for the histogram.

Pixel values below this range will be ignored, and pixel values above this range will be clamped in the sense that they will count towards the highest bin in the histogram. The default value is -8.0..=8.0.

filter: RangeInclusive<f32>

The portion of the histogram to consider when metering.

By default, the darkest 10% and the brightest 10% of samples are ignored, so the default value is 0.10..=0.90.

speed_brighten: f32

The speed at which the exposure adapts from dark to bright scenes, in F-stops per second.

speed_darken: f32

The speed at which the exposure adapts from bright to dark scenes, in F-stops per second.

exponential_transition_distance: f32

The distance in F-stops from the target exposure from where to transition from animating in linear fashion to animating exponentially. This helps against jittering when the target exposure keeps on changing slightly from frame to frame, while still maintaining a relatively slow animation for big changes in scene brightness.

ev
                      â”â—â”
|              â¬ˆ         â”œ exponential section
â”‚        â¬ˆ               â”˜
â”‚    â¬ˆ                   â”
â”‚  â¬ˆ                     â”œ linear section
â”‚â¬ˆ                       â”˜
â—â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€ time

The default value is 1.5.

metering_mask: Handle<Image>

The mask to apply when metering. The mask will cover the entire screen, where:

(0.0, 0.0) is the top-left corner,

(1.0, 1.0) is the bottom-right corner.

Only the red channel of the texture is used. The sample at the current screen position will be used to weight the contribution of each pixel to the histogram:

0.0 means the pixel will not contribute to the histogram,

1.0 means the pixel will contribute fully to the histogram.

The default value is a white image, so all pixels contribute equally.

Usage Notes

The mask is quantized to 16 discrete levels because of limitations in the compute shader implementation.

compensation_curve: Handle<AutoExposureCompensationCurve>

Exposure compensation curve to apply after metering. The default value is a flat line at 0.0. For more information, see AutoExposureCompensationCurve.