Struct bevy_core_pipeline::bloom::BloomSettings

pub struct BloomSettings {
    pub intensity: f32,
    pub low_frequency_boost: f32,
    pub low_frequency_boost_curvature: f32,
    pub high_pass_frequency: f32,
    pub prefilter_settings: BloomPrefilterSettings,
    pub composite_mode: BloomCompositeMode,
}

Applies a bloom effect to an HDR-enabled 2d or 3d camera.

Bloom emulates an effect found in real cameras and the human eye, causing halos to appear around very bright parts of the scene.

See also https://en.wikipedia.org/wiki/Bloom_(shader_effect).

Usage Notes

Bloom is currently not compatible with WebGL2.

Often used in conjunction with bevy_pbr::StandardMaterial::emissive for 3d meshes.

Bloom is best used alongside a tonemapping function that desaturates bright colors, such as crate::tonemapping::Tonemapping::TonyMcMapface.

Bevy’s implementation uses a parametric curve to blend between a set of blurred (lower frequency) images generated from the camera’s view. See https://starlederer.github.io/bloom/ for a visualization of the parametric curve used in Bevy as well as a visualization of the curve’s respective scattering profile.

Fields
intensity: f32

Controls the baseline of how much the image is scattered (default: 0.15).

This parameter should be used only to control the strength of the bloom for the scene as a whole. Increasing it too much will make the scene appear blurry and over-exposed.

To make a mesh glow brighter, rather than increase the bloom intensity, you should increase the mesh’s emissive value.

In energy-conserving mode

The value represents how likely the light is to scatter.

The value should be between 0.0 and 1.0 where:

0.0 means no bloom

1.0 means the light is scattered as much as possible

In additive mode

The value represents how much scattered light is added to the image to create the glow effect.

In this configuration:

0.0 means no bloom

Greater than 0.0 means a proportionate amount of scattered light is added

low_frequency_boost: f32

Low frequency contribution boost. Controls how much more likely the light is to scatter completely sideways (low frequency image).

Comparable to a low shelf boost on an equalizer.

In energy-conserving mode

The value should be between 0.0 and 1.0 where:

0.0 means low frequency light uses base intensity for blend factor calculation

1.0 means low frequency light contributes at full power

In additive mode

The value represents how much scattered light is added to the image to create the glow effect.

In this configuration:

0.0 means no bloom

Greater than 0.0 means a proportionate amount of scattered light is added

low_frequency_boost_curvature: f32

Low frequency contribution boost curve. Controls the curvature of the blend factor function making frequencies next to the lowest ones contribute more.

Somewhat comparable to the Q factor of an equalizer node.

Valid range:

0.0 - base base intensity and boosted intensity are linearly interpolated

1.0 - all frequencies below maximum are at boosted intensity level

high_pass_frequency: f32

Tightens how much the light scatters (default: 1.0).

Valid range:

0.0 - maximum scattering angle is 0 degrees (no scattering)

1.0 - maximum scattering angle is 90 degrees

prefilter_settings: BloomPrefilterSettingscomposite_mode: BloomCompositeMode

Controls whether bloom textures are blended between or added to each other. Useful if image brightening is desired and a must-change if prefilter_settings are used.

Recommendation

Set to BloomCompositeMode::Additive if prefilter_settings are configured in a non-energy-conserving way, otherwise set to BloomCompositeMode::EnergyConserving.

Implementations

impl BloomSettings

pub const NATURAL: Self = _

The default bloom preset.

This uses the EnergyConserving composite mode.

pub const OLD_SCHOOL: Self = _

A preset that’s similar to how older games did bloom.

pub const SCREEN_BLUR: Self = _

A preset that applies a very strong bloom, and blurs the whole screen.