bevy_core_pipeline::bloom
Type Alias BloomSettings

pub type BloomSettings = Bloom;

๐Deprecated since 0.15.0: Renamed to Bloom

Aliased Type

struct BloomSettings {
    pub intensity: f32,
    pub low_frequency_boost: f32,
    pub low_frequency_boost_curvature: f32,
    pub high_pass_frequency: f32,
    pub prefilter: BloomPrefilter,
    pub composite_mode: BloomCompositeMode,
    pub max_mip_dimension: u32,
    pub uv_offset: f32,
}

Fields
intensity: f32

Controls the baseline of how much the image is scattered (default: 0.15).

This parameter should be used only to control the strength of the bloom for the scene as a whole. Increasing it too much will make the scene appear blurry and over-exposed.

To make a mesh glow brighter, rather than increase the bloom intensity, you should increase the meshโs emissive value.

In energy-conserving mode

The value represents how likely the light is to scatter.

The value should be between 0.0 and 1.0 where:

0.0 means no bloom

1.0 means the light is scattered as much as possible

In additive mode

The value represents how much scattered light is added to the image to create the glow effect.

In this configuration:

0.0 means no bloom

Greater than 0.0 means a proportionate amount of scattered light is added

low_frequency_boost: f32

Low frequency contribution boost. Controls how much more likely the light is to scatter completely sideways (low frequency image).

Comparable to a low shelf boost on an equalizer.

In energy-conserving mode

The value should be between 0.0 and 1.0 where:

0.0 means low frequency light uses base intensity for blend factor calculation

1.0 means low frequency light contributes at full power

In additive mode

The value represents how much scattered light is added to the image to create the glow effect.

In this configuration:

0.0 means no bloom

Greater than 0.0 means a proportionate amount of scattered light is added

low_frequency_boost_curvature: f32

Low frequency contribution boost curve. Controls the curvature of the blend factor function making frequencies next to the lowest ones contribute more.

Somewhat comparable to the Q factor of an equalizer node.

Valid range:

0.0 - base intensity and boosted intensity are linearly interpolated

1.0 - all frequencies below maximum are at boosted intensity level

high_pass_frequency: f32

Tightens how much the light scatters (default: 1.0).

Valid range:

0.0 - maximum scattering angle is 0 degrees (no scattering)

1.0 - maximum scattering angle is 90 degrees

prefilter: BloomPrefilter

Controls the threshold filter used for extracting the brightest regions from the input image before blurring them and compositing back onto the original image.

Changing these settings creates a physically inaccurate image and makes it easy to make the final result look worse. However, they can be useful when emulating the 1990s-2000s game look. See BloomPrefilter for more information.

composite_mode: BloomCompositeMode

Controls whether bloom textures are blended between or added to each other. Useful if image brightening is desired and a must-change if prefilter is used.

Recommendation

Set to BloomCompositeMode::Additive if prefilter is configured in a non-energy-conserving way, otherwise set to BloomCompositeMode::EnergyConserving.

max_mip_dimension: u32

Maximum size of each dimension for the largest mipchain texture used in downscaling/upscaling. Only tweak if you are seeing visual artifacts.

uv_offset: f32

UV offset for bloom shader. Ideally close to 2.0 / max_mip_dimension. Only tweak if you are seeing visual artifacts.