bevy_core_pipeline
Module core_2d

Modules

graph

Structs

AlphaMask2d

Alpha mask 2D BinnedPhaseItems.

AlphaMask2dBinKey

Data that must be identical in order to batch phase items together.

Camera2d

A 2D camera component. Enables the 2D render graph for a Camera.

Camera2dBundleDeprecated

Core2dPlugin

MainOpaquePass2dNode

A bevy_render::render_graph::Node that runs the Opaque2d ViewBinnedRenderPhases and AlphaMask2d ViewBinnedRenderPhases

MainTransparentPass2dNode

Opaque2d

Opaque 2D BinnedPhaseItems.

Opaque2dBinKey

Data that must be identical in order to batch phase items together.

Transparent2d

Transparent 2D SortedPhaseItems.

Constants

CORE_2D_DEPTH_FORMAT

Functions

extract_core_2d_camera_phases

prepare_core_2d_depth_textures