Module bevy_core_pipeline::core_3d

Modules

graph

Structs

AlphaMask3d

Camera3d

Configuration for the “main 3d render graph”. The camera coordinate space is right-handed x-right, y-up, z-back. This means “forward” is -Z.

Camera3dBundle

The camera coordinate space is right-handed x-right, y-up, z-back. This means “forward” is -Z.

Camera3dDepthTextureUsage

Core3dPlugin

MainOpaquePass3dNode

A bevy_render::render_graph::Node that runs the Opaque3d [BinnedRenderPhase] and AlphaMask3d bevy_render::render_phase::SortedRenderPhases.

MainTransparentPass3dNode

A bevy_render::render_graph::Node that runs the Transparent3d [SortedRenderPhase].

Opaque3d

Opaque 3D BinnedPhaseItems.

Opaque3dBinKey

Data that must be identical in order to batch phase items together.

Transmissive3d

Transparent3d

ViewTransmissionTexture

Enums

Camera3dDepthLoadOp

The depth clear operation to perform for the main 3d pass.

ScreenSpaceTransmissionQuality

The quality of the screen space transmission blur effect, applied to whatever’s “behind” transmissive objects when their roughness is greater than 0.0.

Constants

CORE_3D_DEPTH_FORMAT

DEPTH_TEXTURE_SAMPLING_SUPPORTED

True if multisampled depth textures are supported on this platform.

Functions

check_msaa

extract_camera_prepass_phase

extract_core_3d_camera_phases

prepare_core_3d_depth_textures

prepare_core_3d_transmission_textures

prepare_prepass_textures