bevy_core_pipeline
Module motion_blur

Per-object, per-pixel motion blur.

Add the MotionBlur component to a camera to enable motion blur.

Modules

node

pipeline

Structs

MotionBlur

A component that enables and configures motion blur when added to a camera.

MotionBlurBundleDeprecated

Adds MotionBlur and the required depth and motion vector prepasses to a camera entity.

MotionBlurPlugin

Adds support for per-object motion blur to the app. See MotionBlur for details.

Constants

MOTION_BLUR_SHADER_HANDLE