bevy_core_pipeline
Module oit

Order Independent Transparency (OIT) for 3d rendering. See OrderIndependentTransparencyPlugin for more details.

Modules

resolve

Module that defines the necesasry systems to resolve the OIT buffer and render it to the screen.

Structs

OitBuffers

Holds the buffers that contain the data of all OIT layers. We use one big buffer for the entire app. Each camaera will reuse it so it will always be the size of the biggest OIT enabled camera.

OrderIndependentTransparencyPlugin

A plugin that adds support for Order Independent Transparency (OIT). This can correctly render some scenes that would otherwise have artifacts due to alpha blending, but uses more memory.

OrderIndependentTransparencySettings

Used to identify which camera will use OIT to render transparent meshes and to configure OIT.

OrderIndependentTransparencySettingsOffset

Constants

OIT_DRAW_SHADER_HANDLE

Shader handle for the shader that draws the transparent meshes to the OIT layers buffer.

Functions

prepare_oit_buffers

This creates or resizes the oit buffers for each camera. It will always create one big buffer that’s as big as the biggest buffer needed. Cameras with smaller viewports or less layers will simply use the big buffer and ignore the rest.