bevy_core_pipeline::oit
Module resolve

Module that defines the necesasry systems to resolve the OIT buffer and render it to the screen.

Modules

node

Contains the render node used to run the resolve pass.

Structs

OitResolveBindGroup

Bind group for the OIT resolve pass.

OitResolvePipeline

Bind group layouts used for the OIT resolve pass.

OitResolvePipelineId

OitResolvePipelineKey

This key is used to cache the pipeline id and to specialize the render pipeline descriptor.

OitResolvePlugin

Plugin needed to resolve the Order Independent Transparency (OIT) buffer to the screen.

Constants

OIT_RESOLVE_SHADER_HANDLE

Shader handle for the shader that sorts the OIT layers, blends the colors based on depth and renders them to the screen.

Functions

prepare_oit_resolve_bind_group

queue_oit_resolve_pipeline