bevy_core_pipeline
Module post_process

Miscellaneous built-in postprocessing effects.

Currently, this consists only of chromatic aberration.

Structs

ChromaticAberration

Adds colored fringes to the edges of objects in the scene.

ChromaticAberrationUniform

The on-GPU version of the ChromaticAberration settings.

PostProcessingNode

The render node that runs the built-in postprocessing stack.

PostProcessingPipeline

GPU pipeline data for the built-in postprocessing stack.

PostProcessingPipelineId

A component attached to cameras in the render world that stores the specialized pipeline ID for the built-in postprocessing stack.

PostProcessingPipelineKey

A key that uniquely identifies a built-in postprocessing pipeline.

PostProcessingPlugin

A plugin that implements a built-in postprocessing stack with some common effects.

PostProcessingUniformBufferOffsets

A component, part of the render world, that stores the appropriate byte offset within the PostProcessingUniformBuffers for the camera it’s attached to.

PostProcessingUniformBuffers

A resource, part of the render world, that stores the ChromaticAberrationUniforms for each view.

Functions

prepare_post_processing_pipelines

Specializes the built-in postprocessing pipeline for each applicable view.

prepare_post_processing_uniforms

Gathers the built-in postprocessing settings for every view and uploads them to the GPU.