bevy_core_pipeline::prepass
Struct ViewPrepassTextures

pub struct ViewPrepassTextures {
    pub depth: Option<ColorAttachment>,
    pub normal: Option<ColorAttachment>,
    pub motion_vectors: Option<ColorAttachment>,
    pub deferred: Option<ColorAttachment>,
    pub deferred_lighting_pass_id: Option<ColorAttachment>,
    pub size: Extent3d,
}

Textures that are written to by the prepass.

This component will only be present if any of the relevant prepass components are also present.

Fields
depth: Option<ColorAttachment>

The depth texture generated by the prepass. Exists only if DepthPrepass is added to the ViewTarget

normal: Option<ColorAttachment>

The normals texture generated by the prepass. Exists only if NormalPrepass is added to the ViewTarget

motion_vectors: Option<ColorAttachment>

The motion vectors texture generated by the prepass. Exists only if MotionVectorPrepass is added to the ViewTarget

deferred: Option<ColorAttachment>

The deferred gbuffer generated by the deferred pass. Exists only if DeferredPrepass is added to the ViewTarget

deferred_lighting_pass_id: Option<ColorAttachment>

A texture that specifies the deferred lighting pass id for a material. Exists only if DeferredPrepass is added to the ViewTarget

size: Extent3d

The size of the textures.

Implementations

impl ViewPrepassTextures

pub fn depth_view(&self) -> Option<&TextureView>

pub fn normal_view(&self) -> Option<&TextureView>

pub fn motion_vectors_view(&self) -> Option<&TextureView>

pub fn deferred_view(&self) -> Option<&TextureView>