bevy_core_pipeline
Module tonemapping

Structs

TonemappingLuts

3D LUT (look up table) textures used for tonemapping

TonemappingNode

TonemappingPipeline

TonemappingPipelineKey

TonemappingPlugin

ViewTonemappingPipeline

Enums

DebandDither

Enables a debanding shader that applies dithering to mitigate color banding in the final image for a given Camera entity.

Tonemapping

Optionally enables a tonemapping shader that attempts to map linear input stimulus into a perceptually uniform image for a given Camera entity.

Functions

get_lut_bind_group_layout_entries

get_lut_bindings

lut_placeholder

prepare_view_tonemapping_pipelines